chore: move to server plugin
